requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
jellyfish==0.7.1
mpmath==1.1.0
numpy==1.16.2
olca-ipc==0.0.7
requests==2.21.0
scipy==1.2.1
sympy==1.4
urllib3==1.24.1
xlrd==1.2.0
